earangodb (eArangoDB v0.1)

This module provides functions to work with collections, documents and graphs. It is a main API for this library.

Functions

collection_create(Name)

Specs

collection_create(Name :: binary()) -> ok | {error, Reason :: term()}.

Creates a new collection with a given name.

collection_delete(Name)

Specs

collection_delete(Name :: binary()) -> ok | {error, Reason :: term()}.

Drops the collection identified by collection-name.

collections_list()

Specs

collections_list() -> [earangodb_colections:collection_info()].

Returns an object with an attribute result containing an array of all collection descriptions.