earangodb (eArangoDB v0.1)

This module provides functions to work with collections, documents and graphs. It is a main API for this library.

Types

ok_response_or_error_reason/0

Specs

ok_response_or_error_reason() :: {ok, map()} | {error, Reason :: term()}.

Functions

collection_create(Name)

Specs

collection_create(Name :: binary()) -> ok_response_or_error_reason().

The same as collection_create(Name, document) function

collection_create(Name, CollectionType)

Specs

collection_create(Name :: binary(), CollectionType :: earangodb_collections:collection_type()) ->
                     ok_response_or_error_reason().

Creates a new collection with a given name and type.

collection_delete(Name)

Specs

collection_delete(Name :: binary()) -> ok_response_or_error_reason().

Drops the collection identified by collection-name.

collections_list()

Specs

collections_list() -> [earangodb_collections:collection_info()].

Returns an object with an attribute result containing an array of all collection descriptions.

document_create(CollectionName, Document)

Specs

document_create(CollectionName :: binary(), Document :: map()) -> ok_response_or_error_reason().

Creates a new document from the document given in the body, unless there is already a document with the _key given.

document_delete(CollectionName, DocumentKey)

Specs

document_delete(CollectionName :: binary(), DocumentKey :: binary()) ->
                   ok_response_or_error_reason().

Deletes the specified document with a given key.

document_read(CollectionName, DocumentKey)

Specs

document_read(CollectionName :: binary(), DocumentKey :: binary()) ->
                 ok_response_or_error_reason().

Returns the document identified by document-id. The returned document contains three special attribute _key containing key which uniquely identifies a document in a given collection.

document_replace(CollectionName, DocumentKey, Document)

Specs

document_replace(CollectionName :: binary(), DocumentKey :: binary(), Document :: map()) ->
                    ok_response_or_error_reason().

Replaces the specified document with the one in the body, provided there is such a document .

document_update(CollectionName, DocumentKey, Document)

Specs

document_update(CollectionName :: binary(), DocumentKey :: binary(), Document :: map()) ->
                   ok_response_or_error_reason().

Partially updates the document identified by document-id. The body of the request must contain a JSON document with the attributes to patch (the patch document). All attributes from the patch document will be added to the existing document if they do not yet exist, and overwritten in the existing document if they do exist there.

graph_edge_create(GraphName, EdgesCollectionName, FromVertex, ToVertex, Properties)

Specs

graph_edge_create(GraphName :: binary(),
                  EdgesCollectionName :: binary(),
                  FromVertex :: integer(),
                  ToVertex :: integer(),
                  Properties :: map()) ->
                     ok_response_or_error_reason().

Creates a new edge in the collection.

graph_edge_delete(GraphName, CollectionName, EdgeKey)

Specs

graph_edge_delete(GraphName :: binary(), CollectionName :: binary(), EdgeKey :: binary()) ->
                     ok_response_or_error_reason().

Removes an edge from the collection.

graph_edge_get(GraphName, CollectionName, EdgeKey)

Specs

graph_edge_get(GraphName :: binary(), CollectionName :: binary(), EdgeKey :: binary()) ->
                  ok_response_or_error_reason().

Gets an edge from the given collection.

graph_edge_update(GraphName, CollectionName, EdgeKey, UpdateDoc)

Specs

graph_edge_update(GraphName :: binary(),
                  CollectionName :: binary(),
                  EdgeKey :: binary(),
                  UpdateDoc :: map()) ->
                     ok_response_or_error_reason().

Updates an edge from the given collection.

graph_vertex_create(GraphName, CollectionName, VertexDocument)

Specs

graph_vertex_create(GraphName :: binary(), CollectionName :: binary(), VertexDocument :: map()) ->
                       ok_response_or_error_reason().

Adds a vertex to the given collection.

graph_vertex_delete(GraphName, CollectionName, VertexKey)

Specs

graph_vertex_delete(GraphName :: binary(), CollectionName :: binary(), VertexKey :: binary()) ->
                       ok_response_or_error_reason().

Removes a vertex from the collection.

graph_vertex_get(GraphName, CollectionName, VertexKey)

Specs

graph_vertex_get(GraphName :: binary(), CollectionName :: binary(), VertexKey :: binary()) ->
                    ok_response_or_error_reason().

Gets a vertex from the collection with a given key.

graph_vertex_replace(GraphName, CollectionName, VertexKey, ReplaceDoc)

Specs

graph_vertex_replace(GraphName :: binary(),
                     CollectionName :: binary(),
                     VertexKey :: binary(),
                     ReplaceDoc :: map()) ->
                        ok_response_or_error_reason().

Replaces the data of the specific vertex in the collection.

graph_vertex_update(GraphName, CollectionName, VertexKey, UpdateDoc)

Specs

graph_vertex_update(GraphName :: binary(),
                    CollectionName :: binary(),
                    VertexKey :: binary(),
                    UpdateDoc :: map()) ->
                       ok_response_or_error_reason().

Updates the data of the specific vertex in the collection.

graphs_create(GraphName, EdgesCollection, FromCollections, ToCollections)

Specs

graphs_create(GraphName :: binary(),
              EdgesCollection :: binary(),
              FromCollections :: [binary()],
              ToCollections :: [binary()]) ->
                 ok_response_or_error_reason().

The creation of a graph requires the name of the graph and a definition of its edges.

graphs_delete(GraphName)

Specs

graphs_delete(GraphName :: binary()) -> ok_response_or_error_reason().

Drops an existing graph object by name.

graphs_get(GraphName)

Specs

graphs_get(GraphName :: binary()) -> ok_response_or_error_reason().

Selects information for a given graph. Will return the edge definitions as well as the orphan collections. Or returns a 404 if the graph does not exist.

graphs_list()

Specs

graphs_list() -> ok_response_or_error_reason().

Lists all graphs stored in this database.